feat: modularize components and integrate virtualized stock table
