+++ package.json
@@ -15,7 +15,8 @@
     "js-cookie": "^3.0.5",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "react-select": "^5.10.2"
+    "react-select": "^5.10.2",
+    "react-window": "^1.8.8"
   },
   "devDependencies": {
     "@babel/core": "^7.24.0",
